Adding Extensive Bas Tab Completion
Tab completion will now use the calling script to generate suggestions.
Because of this those suggestions will work with the alias scheme
described in the README file.

These completion functions will also suggest the commands hcl allows.

When the alias command is in use it will complete the project ids and
task ids with descriptions.

The other commands that use aliases will complete them as before however
they used the calling script to get the list of aliases.  That way the
any bash alises using env's will be called and preserve the
configuration.  This will allow the user to only add one line to their
bashrc/profile w/e and not a line for each  bash alias the user may
need.